requirements.txt
python-dotenv
py-cord
asyncio
g4f
